# deploy/helm/fuzekeys/files/openapi.yaml
...
openapi: 3.1.0
info:
  title: FuzeKeys API
  description: 'FuzeKeys — identity and credential vault. Managed digital identities, per-site accounts,
    encrypted vault assets, capability-token secret brokering, and automated signup workflows.


    DERIVED DOCUMENT: regenerate with `python scripts/export-openapi.py`. Do not hand-edit.'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.0.0
paths:
...
  /api/sms/register-device:
    post:
      tags:
      - SMS
      - SMS
      summary: Register Device
      description: 'Register a new SMS interceptor device.


        SECURITY / BOOTSTRAP TRUST ASSUMPTION: This endpoint is intentionally left

        UNauthenticated because it is the bootstrap that ISSUES the per-device API

        key — requiring the device key here would be a chicken-and-egg problem and

        break first-time registration. The trust assumption is therefore that

        registration is reachable only by intended devices (network/ingress policy

        or an out-of-band enrolment secret should gate it in production). Note it is

        also self-overwriting: re-registering an existing device_id rotates its key,

        so exposure should be limited by the surrounding network controls.

        PRODUCTION NOTE: add an enrolment token / mutual-TLS / signed attestation so

        arbitrary callers cannot register or hijack a device_id.'
      operationId: post_sms_register_device
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRegistrationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sms/otp:
    post:
      tags:
      - SMS
      - SMS
      summary: Receive Otp
      description: 'Receive OTP code from mobile app.


        SECURITY: This endpoint accepts device-submitted data and is therefore

        authenticated. The submitting device must present the API key it was

        issued at registration (via the ``X-Device-Key`` header) and must specify

        which pending request it is fulfilling. The device is then verified to be

        the device that owns/was assigned that request. This closes the OTP

        hijacking hole where any unauthenticated caller could complete the

        "first waiting" request with an attacker-controlled OTP.'
      operationId: post_sms_otp
      parameters:
      - name: X-Device-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sms/requests/{device_id}:
    get:
      tags:
      - SMS
      - SMS
      summary: Get Otp Requests
      description: 'Get pending OTP requests for a device.


        SECURITY: This is a DEVICE-to-server callback (the mobile interceptor polls

        for work it should fulfil), so it is authenticated with the per-device API

        key exactly like ``/otp`` — not the user JWT. The device must present the

        ``X-Device-Key`` it was issued at registration, and that key must belong to

        the ``device_id`` in the path, preventing one device from enumerating

        another device''s pending requests. Unknown/unauthenticated devices get 401.'
      operationId: get_sms_requests_by_device_id
      parameters:
      - name: device_id
        in: path
        required: true
        schema:
          type: string
          title: Device Id
      - name: X-Device-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Device-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sms/request-otp:
    post:
      tags:
      - SMS
      - SMS
      summary: Request Otp
      description: 'Request an OTP for a specific service (called by your main app).


        SECURITY: This is an OPERATOR/USER-facing action (the main app asks the

        platform to wait for an OTP for a given service), so it requires the

        application JWT via get_current_user. An unauthenticated caller could

        otherwise spam OTP requests / push fake jobs to devices. Devices do not

        call this endpoint, so device-key auth is not appropriate here.'
      operationId: post_sms_request_otp
      security:
      - HTTPBearer: []
      parameters:
      - name: service
        in: query
        required: true
        schema:
          type: string
          title: Service
      - name: timeout_seconds
        in: query
        required: false
        schema:
          type: integer
          default: 300
          title: Timeout Seconds
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sms/request-status/{request_id}:
    get:
      tags:
      - SMS
      - SMS
      summary: Get Request Status
      description: 'Get the status of an OTP request.


        SECURITY: This endpoint returns the received OTP value itself

        (``otp_code``), which is highly sensitive. It is operator/user-facing

        (the main app polls for the result), so it requires the application JWT.

        Leaving it unauthenticated would let any caller read OTP codes by guessing

        or enumerating request_ids. Device-key auth is not used because devices

        submit OTPs (via /otp), they do not read them back.'
      operationId: get_sms_request_status_by_request_id
      security:
      - HTTPBearer: []
      parameters:
      - name: request_id
        in: path
        required: true
        schema:
          type: string
          title: Request Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/sms/devices:
    get:
      tags:
      - SMS
      - SMS
      summary: Get Devices
      description: 'Get all registered SMS devices.


        SECURITY: This lists the full device inventory (ids, names, OS/app

        versions, activity) which is sensitive operational/PII-ish data and an

        enumeration aid for attackers. It is an operator/admin view, so it requires

        the application JWT. Not a device callback, so device-key auth does not fit.'
      operationId: get_sms_devices
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    DeviceRegistrationRequest:
      properties:
        device_id:
          type: string
          title: Device Id
        device_name:
          type: string
          title: Device Name
        os_version:
          type: string
          title: Os Version
        app_version:
          type: string
          title: App Version
      additionalProperties: false
      type: object
      required:
      - device_id
      - device_name
      - os_version
      - app_version
      title: DeviceRegistrationRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OtpRequest:
      properties:
        otp:
          type: string
          title: Otp
        sender:
          type: string
          title: Sender
        message_body:
          type: string
          title: Message Body
        timestamp:
          type: integer
          title: Timestamp
        device_id:
          type: string
          title: Device Id
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        request_id:
          type: string
          title: Request Id
      type: object
      required:
      - otp
      - sender
      - message_body
      - timestamp
      - device_id
      - request_id
      title: OtpRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...